+++ R/backend-mssql.R
@@ -372,6 +372,18 @@ simulate_mssql <- function(version = "15.0") {
       get_day = function(x) {
         sql_expr(DATEPART(DAY, !!x))
       },
+      date_count_between = function(start, end, precision, ..., n = 1L){
+
+        check_dots_empty()
+        if (precision != "day") {
+          cli_abort("{.arg precision} must be {.val day} on SQL backends.")
+        }
+        if (n != 1) {
+          cli_abort("{.arg n} must be {.val 1} on SQL backends.")
+        }
+
+        sql_expr(DATEDIFF(DAY, !!start, !!end))
+      },
 
       difftime = function(time1, time2, tz, units = "days") {
 
@@ -383,7 +395,7 @@ simulate_mssql <- function(version = "15.0") {
           cli::cli_abort('The only supported value for {.arg units} on SQL backends is "days"')
         }
 
-        sql_expr(DATEDIFF(DAY, !!time1, !!time2))
+        sql_expr(DATEDIFF(DAY, !!time2, !!time1))
       }
     )