+++ R/backend-spark-sql.R
@@ -44,7 +44,7 @@ simulate_spark_sql <- function() simulate_dbi("Spark SQL")
        },
        add_years = function(x, n, ...) {
          check_dots_empty()
-         sql_expr(add_months(!!!x, !!n*12))
+         sql_expr(add_months(!!x, !!n*12))
        },
        date_build = function(year, month = 1L, day = 1L, ..., invalid = NULL) {
          sql_expr(make_date(!!year, !!month, !!day))
@@ -57,6 +57,18 @@ simulate_spark_sql <- function() simulate_dbi("Spark SQL")
        },
        get_day = function(x) {
          sql_expr(date_part('DAY', !!x))
+       },
+       date_count_between = function(start, end, precision, ..., n = 1L){
+
+         check_dots_empty()
+         if (precision != "day") {
+           cli_abort("{.arg precision} must be {.val day} on SQL backends.")
+         }
+         if (n != 1) {
+           cli_abort("{.arg n} must be {.val 1} on SQL backends.")
+         }
+
+         sql_expr(datediff(!!end, !!start))
        },
 
        difftime = function(time1, time2, tz, units = "days") {
@@ -126,12 +138,8 @@ simulate_spark_sql <- function() simulate_dbi("Spark SQL")
                                    analyze = TRUE,
                                    in_transaction = FALSE) {
 
-  if (temporary) {
-    sql <- sql_values_subquery(con, values, types = types, lvl = 1)
-    db_compute(con, table, sql, overwrite = overwrite)
-  } else {
-    NextMethod()
-  }
+  sql <- sql_values_subquery(con, values, types = types, lvl = 1)
+  db_compute(con, table, sql, overwrite = overwrite, temporary = temporary)
 }
 
 #' @export
@@ -146,14 +154,11 @@ simulate_spark_sql <- function() simulate_dbi("Spark SQL")
                                    analyze = TRUE,
                                    in_transaction = FALSE) {
 
-  if (!temporary) {
-    cli::cli_abort("Spark SQL only support temporary tables")
-  }
-
   sql <- glue_sql2(
     con,
     "CREATE ", if (overwrite) "OR REPLACE ",
-    "TEMPORARY VIEW {.tbl {table}} AS \n",
+    if (temporary) "TEMPORARY VIEW" else "TABLE",
+    " {.tbl {table}} AS \n",
     "{.from {sql}}"
   )
   DBI::dbExecute(con, sql)